+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog ##
 #### dev ####
 * Make `Loader.ARGS` public so mods can check program arguments
+* Optionally include LWJGL3 for Dear ImGui usage
 
 #### v3.29.0 ####
 * Add clear error if patch method is non-static